Rename std-lib directory to workspace
