+++ libs/language-server/.gitignore
@@ -4,4 +4,4 @@
 
 # files generated by Langium
 src/lib/ast/generated
-src/lib/builtin-library/generated
+src/lib/workspace/generated
