+++ libs/language-server/src/lib/jayvee-module.ts
@@ -24,7 +24,6 @@ import {
 } from './ast/generated/module';
 import { ValueTypeProvider } from './ast/wrappers/value-type/primitive/primitive-value-type-provider';
 import { WrapperFactoryProvider } from './ast/wrappers/wrapper-factory-provider';
-import { JayveeWorkspaceManager } from './builtin-library/jayvee-workspace-manager';
 import { JayveeValueConverter } from './jayvee-value-converter';
 import {
   JayveeCodeActionProvider,
@@ -37,6 +36,7 @@ import {
 import { JayveeImportResolver } from './services/import-resolver';
 import { RuntimeParameterProvider } from './services/runtime-parameter-provider';
 import { JayveeValidationRegistry } from './validation/validation-registry';
+import { JayveeWorkspaceManager } from './workspace/jayvee-workspace-manager';
 
 /**
  * Declaration of custom services for the Jayvee language.
