+++ libs/language-server/src/lib/workspace/index.ts
