+++ libs/language-server/src/lib/workspace/jayvee-workspace-manager.ts
